@@ -10,30 +10,29 @@
   },
   "dependencies": {
     "@stitches/react": "^1.2.1",
+    "browserslist": "^4.16.5",
     "date-fns": "^2.25.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0"
+    "immer": "^9.0.6",
+    "nth-check": "^2.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.0.0"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-router-dom": "^5.3.2",
+    "@testing-library/react": "^13.0.0",
+    "@testing-library/user-event": "^14.0.0",
+    "@types/jest": "^29.0.0",
+    "@types/node": "^18.0.0",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+    "@types/react-router-dom": "^5.3.3",
     "prettier": "^2.4.1",
-    "react-scripts": "^4.0.3",
+    "react-scripts": "^5.0.1",
     "typescript": "^4.5.2"
   },
   "resolutions": {
-    "react-scripts/**/ansi-regex": "^5.0.1",
-    "react-scripts/**/nth-check": "^2.0.1",
-    "react-scripts/**/immer": "^9.0.6",
-    "react-scripts/**/glob-parent": "^5.1.2",
-    "react-scripts/**/browserslist": "^4.16.5"
+    "react-scripts/**/nth-check": "^2.0.1"
   },
   "eslintConfig": {
     "extends": [
